Define a recursive schema for ActiveConstraints
MUIR-25

This is a proposal based on discussions captured in MUIR-25 for a
recursive structure for ActiveConstraints POSTed to /search. See the
example for intended usage. Basically, you can nest AND, OR, and NOT
queries within a single facet.

Note that if we did want to open this up to allow nesting of additional
facet names, for truly complex queries, it would be as simple as
allowing constraintNames in addition to 'and', 'or', and 'not' - but
absent a requirement, I think that is more burden on the middle-tier
implementation than we need.

Geert, you had proposed other kinds of facets ... would they be covered
by this structure? (I can see, for example, that this misses less-than
and greater-than.) Would that be an extension of this spec?

I'm not intending this commit as a complete proposal, but I wanted to
try mapping out try recursive schema. I also don't think it would be
hard to implement on the middle-tier, so I don't think it is necessary
to limit levels of recursion.