Add TypedFilters
TypedFilters are specific types that must be understood by the
middle-tier. Once you enter a TypedFilter, you are no longer within a
recursive CombinedFilter. They are leaf filters within the recursive
CombinedFilter structure and will have special handlers in the middle
tier.

Also, checking in the equivalent yaml file, for ease of development.