Flesh out SearchResponse
Note that this very closely matches what comes from the MarkLogic Search
API. Most of the items are optional, however.

Note that with SearchResults, I specified 'uri' instead of 'id', for the
moment. At the moment, I am not clear on how we would map to an id,
which I assume would combine with a 'type' understood by the middle tier
and map eventually to a URI. But this assumes that each document is
associated with an explicit type, which will not be the case when
searching arbitrary documents in a database.

Perhaps we need to support two cases: the preferred case where each
result has a supported type and an ID, and a fallback where it only has
a URI.

MUIR-25