GROVE-187: Use same Docker container for all tests
This allows the tests to run against a consistent MarkLogic. It
currently requires you to have a local image named
`grove-marklogic:1.0.0-rc.1-mlDeployed-9.0-8`

A next step will be to host an image in an externally-useful place.

I experimented with allowing more than one Docker container (per test
file), but they collided with each other on port 8063. A way around this
might be to require that the middle-tier also be run inside Docker, and
actually spin up different versions of the middle-tier itself (connected
to MarkLogic on random ports).

So, for now, the entire test suite shares the database and tests have to
be careful not to pollute it.